pyproject.toml
[project]
name = "tu-hogar-whatsapp-reminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.168.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pytest>=8.3.5",
    "resend>=2.7.0",
    "ruff>=0.11.7",
    "twilio>=9.5.2",
]
